Add a userreq function to udp and actually start doing some stuff.
This creates a socket and does the start of the bind bits, but most
importantly it calls all the correct things and does it in the correct
manner. Now we know it works it's  acase of beefing it out.

Next is probably sendto/recvfrom to see if we can actually send via
the code we have.