In the brave new world these functions are really "core" functions and so i've moved them here where they make more sense. Also this allows us to build the kernel module for "core".
The old ones will be removed from the util directory shortly.

NB, this commit of pools.c has the fix that allows us to work! If we're in
the kernel we need to use B_ANY_KERNEL_ADDRESS when creating areas, not the
vanilla userland B_ANY_ADDRESS as we were using. Tsch, tsch. Also take the
opportunity to add the B_CONTIGUOUS flag to the area creations which should
probably always have been there.