Wow!
These files come from OpenBSD's libc implementation and have been
adapted where needed to work on beos, though they will need more work
and some don't really do what was intended yet! This is intended to
start providing a lot of the functionality that libnet.so on R5 does
and gets us a long way towards it.

Some of this can possibly be simplified and modified, and the warnings
can be removed. If you want to do this, please do!

We really need a /dev/random driver as well as presently the random number
generator is weak and thus not a good thing.