Change some comments and remove some Codewarrior nonsense.
Fix poll so it builds.

I think this poll.c should be removed as it's simply a wrapper around
select. If a lib wants to do poll and use a wrapper then it'll be just as
fast and so I don't really see the benefit of this being here. If we're going
to add poll support we should do it properly and do it in kernel where
it'll be useful.

I think I've said this before... <sigh>