Final part of the accept fix. What was it?
Basically it was an old problem making a return visit. On beos malloc
returns memory with all sorts of values, so when comparing 12 bytes of
a sockaddr when only the first 4 are set for sure we now set the entire
structure to 0 before setting any values, thus ensuring that matching works
correctly. The problem with accept was that we didn't match the arp entry
due to these extra 8 bytes being random values, and so tried to resolve
an entry we already had, which never worked as we just kept updating
the original cache entry!

With this fix in place we're in better shape for the future and also
we can now use R5's ftp with the new stack!