** Warning will Robinson - untested code!! **
I was going to test it, but then the issue with the driver happenned :(

Anyway, this should fix the KDL I saw.
Also, we can't assume that we have a valid reason for reporting an extra
fd available unless we're also waiting on it. hence, if we get a hit for
write but we haven't got a write set (as happens in the test that threw me
into KDL, we can't really report any fd's being ready. Thus I've moved where
we increment the counter to a better place.