This really is horrible, but it works :)
Basically, we can't simply refernce the selectsync being passed in as it's
a userland memory space, and allocated off the heap at that. so, we need our
own copy and then we need to copy it back when we're done. That's what this
does. It's possible to refine it a bit I'm sure.

This now exposes the problem in libnet even more :)