The changes required to build a bootable kernel.
Please note, that while it's my name on the commit the glory
really belongs to Ingo for his support and ideas about how to do
the things I've been asking of him. Ryan also deserves some credit as he
gave me the idea for the final step to getting the config file created.

This really does prove how powerful a tool IRC is :)