Change the prototype for read/write.
NB There is a disconnecte here as Geist points out. We return the amount
of data read/written but we allow the user to pass in a size_t for how
much data is available to read/write, but we return an ssize_t which
can't actually contain as large a number as a size_t, so theoretically
we have a brakage. However, this is how Be had it and so as we're moving
to their semantics and I doubt anyone will be reading/writing more than
ssize_t bytes (~2Gb) anyways...