Changes, changes, changes...
Change the way that device drivers read/write calls return their information.
Basically they now return
0 == success
<0 == error code
The length of data read/written is stored as *len. NB this does require
quite a few changes but should be more flexible and is required
as the beos binary modules will expect this to be the case.